Changes required to fix fast restart
