+++ src/editor/extended_python_generator.ts
@@ -270,7 +270,7 @@ export class ExtendedPythonGenerator extends PythonGenerator {
         }
       });
       if (exported) {
-        const variableName = variableModel.getName();
+        const variableName = variableModel.name;
         const actualVariableName = super.getVariableName(variableModel.getId());
         const getPythonModuleVariableBlock = {
           'kind': 'block',
