+++ src/reactComponents/ModuleNameComponent.tsx
@@ -18,7 +18,8 @@
 /**
  * @author alan@porpoiseful.com (Alan Smith)
  */
-import {TabType} from './Tabs';
+import {TabType } from '../types/TabType';
+
 import * as Antd from 'antd';
 import * as I18Next from 'react-i18next';
 import * as React from 'react';
